+++ FeralCommon/Config/BoolConfig.cs
@@ -1,0 +1,6 @@
+using System.Diagnostics.CodeAnalysis;
+
+namespace FeralCommon.Config;
+
+[SuppressMessage("ReSharper", "ClassNeverInstantiated.Global")]
+public class BoolConfig : Config<BoolConfig, bool>;