+++ FeralCommon/Config/ColorConfig.cs
@@ -1,18 +1,19 @@
+using System.Diagnostics.CodeAnalysis;
 using UnityEngine;
 
 namespace FeralCommon.Config;
 
-public class ColorConfig : StringConfig
+[SuppressMessage("ReSharper", "ClassNeverInstantiated.Global")]
+[SuppressMessage("ReSharper", "UnusedMember.Global")]
+public class ColorConfig : Config<ColorConfig, string>
 {
-    public ColorConfig()
-    {
-        WithCharacterLimit(7);
-        WithNumberOfLines(1);
-        WithTrimText(true);
-    }
-
     public Color ToColor()
     {
         return ColorUtility.TryParseHtmlString(this, out var color) ? color : Color.white;
+    }
+
+    public static implicit operator Color(ColorConfig config)
+    {
+        return config.ToColor();
     }
 }