+++ FeralCommon/Input/IActionContainer.cs
@@ -1,0 +1,8 @@
+using UnityEngine.InputSystem;
+
+namespace FeralCommon.Input;
+
+public interface IActionContainer
+{
+    public InputAction GetAction(string actionId);
+}