+++ FeralCommon/Input/IUnityInput.cs
@@ -1,0 +1,8 @@
+using UnityEngine.InputSystem;
+
+namespace FeralCommon.Input;
+
+public interface IUnityInput
+{
+    Key Key { get; }
+}