+++ FeralCommon/Integrate/LethalConfig.cs
@@ -18,16 +18,26 @@ public static class LethalConfig
     {
         return raw switch
         {
+            BoolConfig config => ToConfigItem(config),
             IntConfig config => ToConfigItem(config),
             FloatConfig config => ToConfigItem(config),
             StringConfig config => ToConfigItem(config),
+            ColorConfig config => ToConfigItem(config),
             _ => throw new ArgumentException($"Unsupported config type: {raw.GetType().Name}")
         };
     }
 
+    private static BoolCheckBoxConfigItem ToConfigItem(BoolConfig config)
+    {
+        return new BoolCheckBoxConfigItem(config, new BoolCheckBoxOptions
+        {
+            RequiresRestart = config.RequiresRestart
+        });
+    }
+
     private static IntSliderConfigItem ToConfigItem(IntConfig config)
     {
-        return new IntSliderConfigItem(config.ConfigEntry, new IntSliderOptions
+        return new IntSliderConfigItem(config, new IntSliderOptions
         {
             Min = config.Min,
             Max = config.Max,
@@ -37,7 +47,7 @@ public static class LethalConfig
 
     private static FloatSliderConfigItem ToConfigItem(FloatConfig config)
     {
-        return new FloatSliderConfigItem(config.ConfigEntry, new FloatStepSliderOptions
+        return new FloatSliderConfigItem(config, new FloatStepSliderOptions
         {
             Min = config.Min,
             Max = config.Max,
@@ -48,12 +58,23 @@ public static class LethalConfig
 
     private static TextInputFieldConfigItem ToConfigItem(StringConfig config)
     {
-        return new TextInputFieldConfigItem(config.ConfigEntry, new TextInputFieldOptions
+        return new TextInputFieldConfigItem(config, new TextInputFieldOptions
         {
             RequiresRestart = config.RequiresRestart,
             CharacterLimit = config.CharacterLimit,
             NumberOfLines = config.NumberOfLines,
             TrimText = config.TrimText
+        });
+    }
+
+    private static TextInputFieldConfigItem ToConfigItem(ColorConfig config)
+    {
+        return new TextInputFieldConfigItem(config, new TextInputFieldOptions
+        {
+            RequiresRestart = config.RequiresRestart,
+            CharacterLimit = 7,
+            NumberOfLines = 1,
+            TrimText = true
         });
     }
 }